move tests to kube2e/gloo
