+++ test/kube2e/gloo/generated_kube_types_test.go
@@ -1,4 +1,4 @@
-package kube_test
+package gloo_test
 
 import (
 	"context"
