+++ test/kube2e/gloo/gloo_suite_test.go
@@ -2,6 +2,11 @@ package gloo_test
 
 import (
 	"context"
+	gatewayv1 "github.com/solo-io/gloo/projects/gateway/pkg/api/v1"
+	gloov1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
+	"github.com/solo-io/k8s-utils/testutils/clusterlock"
+	apiext "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
+	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"os"
 	"path/filepath"
 	"testing"
@@ -41,11 +46,38 @@ var (
 
 	ctx    context.Context
 	cancel context.CancelFunc
+
+	apiExts apiext.Interface
+	locker  *clusterlock.TestClusterLocker
 )
 
-var _ = BeforeSuite(func() {
+var _ = SynchronizedBeforeSuite(beforeSuiteOne, beforeSuiteAll)
+var _ = SynchronizedAfterSuite(afterSuiteOne, afterSuiteAll)
+
+func beforeSuiteOne() []byte {
+	// Register the CRDs once at the beginning of the suite
 	ctx, cancel = context.WithCancel(context.Background())
+	cfg, err := kubeutils.GetConfig("", "")
+	Expect(err).NotTo(HaveOccurred())
+
+	apiExts, err = apiext.NewForConfig(cfg)
+	Expect(err).NotTo(HaveOccurred())
+
+	err = skhelpers.AddAndRegisterCrd(ctx, gloov1.UpstreamCrd, apiExts)
+	Expect(err).NotTo(HaveOccurred())
+
+	err = skhelpers.AddAndRegisterCrd(ctx, gatewayv1.VirtualServiceCrd, apiExts)
+	Expect(err).NotTo(HaveOccurred())
+	return nil
+}
+
+func beforeSuiteAll(_ []byte) {
 	var err error
+	locker, err = clusterlock.NewTestClusterLocker(kube2e.MustKubeClient(), clusterlock.Options{})
+	Expect(err).NotTo(HaveOccurred())
+	Expect(locker.AcquireLock(retry.Attempts(40))).NotTo(HaveOccurred())
+
+	ctx, cancel = context.WithCancel(context.Background())
 	testHelper, err = kube2e.GetTestHelper(ctx, namespace)
 	Expect(err).NotTo(HaveOccurred())
 	skhelpers.RegisterPreFailHandler(helpers.KubeDumpOnFail(GinkgoWriter, testHelper.InstallNamespace))
@@ -62,14 +94,24 @@ var _ = BeforeSuite(func() {
 	Expect(err).NotTo(HaveOccurred())
 
 	snapshotWriter = helpers.NewSnapshotWriter(resourceClientset, []retry.Option{})
-})
+}
 
-var _ = AfterSuite(func() {
+func afterSuiteOne(ctx context.Context) {
+	// Delete those CRDs once at the end of the suite
+	_ = apiExts.ApiextensionsV1().CustomResourceDefinitions().Delete(ctx, gloov1.UpstreamCrd.FullName(), v1.DeleteOptions{})
+	_ = apiExts.ApiextensionsV1().CustomResourceDefinitions().Delete(ctx, gatewayv1.VirtualServiceCrd.FullName(), v1.DeleteOptions{})
+
+	cancel()
+}
+
+func afterSuiteAll(_ context.Context) {
+	err := locker.ReleaseLock()
+	Expect(err).NotTo(HaveOccurred())
+
 	if glootestutils.ShouldTearDown() {
 		uninstallGloo()
 	}
-	cancel()
-})
+}
 
 func installGloo() {
 	cwd, err := os.Getwd()
